🚚 [rendering] Move module `adapters.{cookiecutter.extensions => cookiecutterextensions}`
